removed dsl module
removed the dsl module as it does not really help that much and moved
the stone related functions into the stone module
