--- test/quarto_test/dsl_test.exs
@@ -1,4 +1,0 @@
-defmodule QuartoTest.DSL do
-  use ExUnit.Case, async: true
-  doctest Quarto.DSL
-end
